ColbertReranker must extend the class BaseDocumentCompressor to compress documents
