+++ server/requirements.txt
@@ -38,7 +38,6 @@ accelerate==0.34.0
 pydantic==2.7.4
 pydantic_core==2.18.4
 psycopg2-binary==2.9.9
-colbert-ai
 llama-index
 llama-index-core
 llama-index-postprocessor-colbert-rerank
